AGENT-131: Move ignition and iso to asset/agent/image
