+++ pkg/asset/agent/image/ignition_test.go
@@ -1,4 +1,4 @@
-package agent
+package image
 
 import (
 	"testing"
